Move package node_agent_k8s/nodeagentmgmt to node_agent/management.
