+++ security/cmd/node_agent/management/client.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package nodeagentmgmt
+package management
 
 import (
 	"net"
